pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "securegenomics"
version = "0.1.0"
description = "SecureGenomics CLI - Core logic engine for secure genomic computation"
authors = [{name = "SecureGenomics Team"}]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "click>=8.0.0",
    "requests>=2.28.0",
    "cryptography>=3.4.0",
    "pydantic>=1.10.0",
    "rich>=12.0.0",
    "aiohttp>=3.8.0",
    "asyncio-throttle>=1.0.0",
    "pyjwt>=2.4.0",
    "typer>=0.7.0",
    "pyyaml>=6.0.0",
    "psutil>=5.9.0",
    "python-decouple>=3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.991",
    "pytest-cov>=4.0.0",
]
fhe = [
    "tenseal>=0.3.0",  # For FHE operations
    "numpy>=1.21.0",
]

[project.scripts]
securegenomics = "securegenomics.cli:main"

[project.urls]
Homepage = "https://github.com/securegenomics/secure-genomics-v2"
Documentation = "https://github.com/securegenomics/secure-genomics-v2/docs"
Repository = "https://github.com/securegenomics/secure-genomics-v2"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 